# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Studio Guest Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /preregistrations:
    get:
      operationId: listPreregistrations
      tags: [preregistrations]
      summary: List pre-registered guests
      parameters:
        - name: date
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Filter by date (YYYY-MM-DD), defaults to today
      responses:
        "200":
          description: Pre-registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Preregistration"
    post:
      operationId: createPreregistration
      tags: [preregistrations]
      summary: Pre-register an expected guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreregistrationInput"
      responses:
        "201":
          description: Pre-registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preregistration"
  /preregistrations/{id}:
    delete:
      operationId: deletePreregistration
      tags: [preregistrations]
      summary: Delete a pre-registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /preregistrations/{id}/convert:
    post:
      operationId: convertPreregistration
      tags: [preregistrations]
      summary: Convert a pre-registration to an active check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertPreregistrationInput"
      responses:
        "201":
          description: Guest checked in from pre-registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /public/preregistrations:
    post:
      operationId: createPublicPreregistration
      tags: [preregistrations]
      summary: Submit a visitor self-service pre-registration (no authentication)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicPreregistrationInput"
      responses:
        "201":
          description: Pre-registration submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preregistration"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all system users (admin only)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new system user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user profile and role
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user's display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{clerkId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{clerkId}/password:
    patch:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
...
components:
  schemas:
...
    Guest:
      type: object
      required:
        - id
        - name
        - company
        - hostName
        - purposeOfVisit
        - site
        - status
        - checkinAt
        - badgeId
      properties:
        id:
          type: integer
        badgeId:
          type: string
        name:
          type: string
        company:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hostName:
          type: string
        purposeOfVisit:
          type: string
        site:
          type: string
          description: "Dallas/The Plex, Tustin, Nashville"
        status:
          type: string
          enum: [active, checked_out]
        checkinAt:
          type: string
          format: date-time
        checkoutAt:
          type: ["string", "null"]
          format: date-time
        expectedDeparture:
          type: ["string", "null"]
          format: date-time
        photoUrl:
          type: ["string", "null"]
        checkedInByClerkId:
          type: ["string", "null"]
        checkedOutByClerkId:
          type: ["string", "null"]
        preregistrationId:
          type: ["integer", "null"]
        studios:
          type: array
          items:
            type: string
        timeOnSiteMinutes:
          type: ["integer", "null"]
        isOverdue:
          type: boolean
...
    Preregistration:
      type: object
      required:
        - id
        - guestName
        - company
        - hostName
        - expectedArrival
        - site
        - status
      properties:
        id:
          type: integer
        guestName:
          type: string
        company:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hostName:
          type: string
        purposeOfVisit:
          type: ["string", "null"]
        site:
          type: string
        expectedArrival:
          type: string
          format: date-time
        expectedDeparture:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, converted, cancelled]
        createdByClerkId:
          type: ["string", "null"]
        convertedGuestId:
          type: ["integer", "null"]
        studios:
          type: array
          items:
            type: string
    PreregistrationInput:
      type: object
      required:
        - guestName
        - company
        - hostName
        - expectedArrival
        - site
      properties:
        guestName:
          type: string
          minLength: 1
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        hostName:
          type: string
          minLength: 1
        purposeOfVisit:
          type: string
        site:
          type: string
          minLength: 1
        expectedArrival:
          type: string
          format: date-time
        expectedDeparture:
          type: string
          format: date-time
        studios:
          type: array
          items:
            type: string
    ConvertPreregistrationInput:
      type: object
      properties:
        photoUrl:
          type: string
...
    AppUser:
      type: object
      required:
        - clerkId
        - role
        - createdAt
      properties:
        clerkId:
          type: string
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [security, admin, kiosk]
        createdAt:
          type: string
          format: date-time
    AppUserInput:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
          format: email
          minLength: 3
          description: Login email; must be unique (case-insensitive)
        password:
          type: string
          minLength: 8
          description: Initial password for the operator
        displayName:
          type: string
        role:
          type: string
          enum: [security, admin, kiosk]
    ResetPasswordInput:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 8
    AppUserUpdate:
      type: object
      properties:
        displayName:
          type: string
    RoleUpdate:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [security, admin, kiosk]
...
    PublicPreregistrationInput:
      type: object
      required:
        - guestName
        - hostName
        - site
        - expectedArrival
      properties:
        guestName:
          type: string
          minLength: 1
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        hostName:
          type: string
          minLength: 1
        purposeOfVisit:
          type: string
        site:
          type: string
          minLength: 1
        expectedArrival:
          type: string
          format: date-time
        expectedDeparture:
          type: string
          format: date-time
        studios:
          type: array
          items:
            type: string